+++ haskellopt_gen2/IterCont.pass-0000.opt.hs
@@ -19,8 +19,8 @@ import GHC.Num
 import GHC.Types
 
 nats1 = 
-  let rec st = st : (rec (st + 1)) in
-  (rec 0)
+  let rec st = st : (rec (st + (1::Int))) in
+  (rec (0::Int))
 
 loop1 = \f -> \state -> 
         let λ = \st -> f λ st in
